+++ packages/svelte/tests/compiler-errors/samples/css-global-modifier-start-1/_config.js
@@ -1,0 +1,9 @@
+import { test } from '../../test';
+
+export default test({
+	error: {
+		code: 'css_global_block_invalid_modifier_start',
+		message: 'A `:global` selector can only be modified if it is a descendant of other selectors',
+		position: [75, 77]
+	}
+});