+++ packages/svelte/tests/compiler-errors/samples/css-global-modifier/_config.js
@@ -3,7 +3,7 @@ import { test } from '../../test';
 export default test({
 	error: {
 		code: 'css_global_block_invalid_modifier',
-		message: 'A :global {...} block cannot modify an existing selector',
-		position: [14, 21]
+		message: 'A `:global` selector cannot modify an existing selector',
+		position: [70, 77]
 	}
 });