+++ packages/svelte/tests/runtime-runes/samples/store-assignments/_config.js
@@ -1,6 +1,5 @@
 import { test } from '../../test';
 
 export default test({
-	// TODO JSDOM doesn't support container queries
-	skip: true
+	html: '<p>3</p>'
 });